Rename fizzbuzz -> fizz_buzz everywhere
